Move a few more models into Database
